pyproject.toml
[tool.poetry]
name = "maistro"
version = "0.1.0"
description = "A lightweight agent framework for AI musicians"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "maistro"}]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.0"
prompt-toolkit = "^3.0.43"
pypdf = "^4.1.0"
anthropic = "^0.45.2"
chromadb = "^0.6.3"
requests = {extras = ["socks"], version = "^2.32.3"}
datetime = "^5.5"
google-api-python-client = "^2.161.0"
spotipy = "^2.25.0"
aiohttp = "^3.11.12"
pyotp = "^2.9.0"
google-auth-oauthlib = "^1.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
